Update Makefile to make it competent

+++ Makefile
@@ -5,14 +5,13 @@ otp-gui: src/otp.cpp src/chacha.cpp src/gui/*.cpp
 	g++ src/gui/*.cpp src/otp.cpp src/chacha.cpp -o otp-gui $(shell pkg-config --cflags --libs gtkmm-4.0)
 
 install:
-	cp ./otp /bin/
-	cp ./otp-gui /bin/otp-gui
-	cp src/gui/otp-gui-global.desktop /usr/share/applications/otp-gui.desktop
+	if [ -r ./otp ]; then install ./otp /usr/local/bin/; fi
+	if [ -r ./otp-gui ]; then install ./otp-gui /usr/local/bin/otp-gui; install src/gui/otp-gui-global.desktop /usr/share/applications/otp-gui.desktop; fi
 
 local:
-	cp ./otp ~/.local/bin/
-	cp ./otp-gui ~/.local/bin/otp-gui
+	if [ -r ./otp ]; then install ./otp ~/.local/bin/; fi
+	if [ -r ./otp-gui ]; then install ./otp-gui ~/.local/bin/otp-gui; fi
 
 clean:
-	rm otp
-	rm otp-gui
+	if [ -r ./otp ]; then rm otp; fi
+	if [ -r ./otp-gui ]; then rm otp-gui; fi
